feat: break up scrapeData
