+++ lib/transform.js
@@ -5,31 +5,6 @@ import * as datetime from './datetime.js';
 
 const usStates = JSON.parse(fs.readFileSync(path.join('coronavirus-data-sources', 'lib', 'us-states.json'), 'utf8'));
 
-const countryCodes = JSON.parse(fs.readFileSync(path.join('coronavirus-data-sources', 'ISO-3166-Countries-with-Regional-Codes', 'slim-3', 'slim-3.json'), 'utf8'));
-
-/*
-  Override some incorrect country names
-*/
-const countryMap = {
-  Kosovo: 'XKX',
-  'Congo (Kinshasa)': 'Congo, Democratic Republic of the',
-  "Cote d'Ivoire": "Côte d'Ivoire",
-  Russia: 'Russian Federation',
-  Vietnam: 'Viet Nam',
-  'Korea, South': 'Korea, Republic of',
-  'South Korea': 'Korea, Republic of',
-  'North Korea': "Korea (Democratic People's Republic of)",
-  Brunei: 'Brunei Darussalam',
-  Reunion: 'Réunion',
-  Curacao: 'Curaçao',
-  'United Kingdom': 'GBR',
-  'occupied Palestinian territory': 'PSE',
-  'Congo (Brazzaville)': 'COG',
-  Tanzania: 'TZA',
-  'The Bahamas': 'BHS',
-  'Gambia, The': 'GMB'
-};
-
 /*
   Convert an object keyed on county name to an array
 */
@@ -87,37 +62,6 @@ function hash(string) {
     .createHash('md5')
     .update(string)
     .digest('hex');
-}
-
-/*
-  Normalize the state as a 2-letter string
-*/
-function toUSStateAbbreviation(string) {
-  return usStates[string] || string;
-}
-
-/*
-  Normalize the state as a 2-letter string
-*/
-function toISO3166Alpha3(string) {
-  string = countryMap[string] || string;
-  for (const country of countryCodes) {
-    if (country['alpha-3'] === string || country['alpha-2'] === string || country.name === string || country.name.replace(/\s*\(.*?\)/, '') === string || country.name.replace(/, Province of .*$/, '') === string || country.name.replace(/, Republic of$/, '') === string) {
-      return country['alpha-3'];
-    }
-  }
-  console.warn('⚠️  Could not find country code for', string);
-  return string;
-}
-
-/*
-  Calculates active cases from location data
- */
-function getActiveFromLocation(location) {
-  const cases = location.cases !== undefined ? location.cases : 0;
-  const deaths = location.deaths !== undefined ? location.deaths : 0;
-  const recovered = location.recovered !== undefined ? location.recovered : 0;
-  return cases - deaths - recovered;
 }
 
 /*
@@ -193,13 +137,6 @@ function sumData(dataArray, object) {
 }
 
 /*
-  Get the priority of a location
-*/
-function getPriority(location) {
-  return location.priority !== undefined ? location.priority : 0;
-}
-
-/*
   Get the growth factor for two numbers, null if infinite
 */
 function getGrowthfactor(casesToday, casesYesterday) {
@@ -232,4 +169,4 @@ function addEmptyRegions(regionDataArray, regionNameArray, regionGranularity) {
   return regionDataArray;
 }
 
-export { objectToArray, addCounty, getName, hash, toUSStateAbbreviation, getActiveFromLocation, transposeTimeseries, toISO3166Alpha3, sumData, usStates, getPriority, getGrowthfactor, addEmptyRegions };
+export { objectToArray, addCounty, getName, hash, transposeTimeseries, sumData, usStates, getGrowthfactor, addEmptyRegions };
