+++ tasks/scrapeData/index.js
@@ -1,0 +1,14 @@
+import runScrapers from './runScraper.js';
+import normalizeLocations from './normalizeLocations.js';
+import cleanLocations from './cleanLocations.js';
+import dedupLocations from './dedupData.js';
+import rateLocations from './rateLocations.js';
+
+const scrapeData = async args =>
+  runScrapers(args)
+    .then(normalizeLocations)
+    .then(dedupLocations)
+    .then(rateLocations)
+    .then(cleanLocations);
+
+export default scrapeData;
