+++ tasks/writeData.js
@@ -1,8 +1,7 @@
 import path from 'path';
 import * as fs from '../lib/fs.js';
-import * as stringify from '../lib/stringify.js';
 
-const writeData = async ({ locations, featureCollection, report, options, sourceRatings }) => {
+const writeData = async ({ locations, report, sourceRatings, options }) => {
   let suffix = '';
   if (options.outputSuffix !== undefined) {
     suffix = options.outputSuffix;
@@ -12,15 +11,11 @@ const writeData = async ({ locations, featureCollection, report, options, source
 
   await fs.writeFile(path.join('dist', `data${suffix}.json`), JSON.stringify(locations, null, 2));
 
-  await fs.writeCSV(path.join('dist', `data${suffix}.csv`), stringify.csvForDay(locations));
-
-  await fs.writeJSON(path.join('dist', `features${suffix}.json`), featureCollection);
+  await fs.writeJSON('dist/ratings.json', sourceRatings);
 
   await fs.writeJSON('dist/report.json', report);
 
-  await fs.writeJSON('dist/ratings.json', sourceRatings);
-
-  return { locations, featureCollection, report, options };
+  return { locations, report, options };
 };
 
 export default writeData;
